feat: move credit signals to handlers
